+++ docs/src/markdown/about/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.11
+
+-   **NEW**: Add new command line option `--skip-dict-compile` which will skip the dictionary compiling step if the
+    dictionary already exists. Changes to a custom dictionary will be ignored.
+
 ## 2.10
 
 -   **NEW**: Allow specifying parallel processes to speed up spell checking. Number of jobs can be specified either